合并 main 累积改动（fund AUM dashboard / migration 046-049 / security
C-1 / db-promote 工具等），解决 3 处冲突并完成 PR #27 review 修复。

冲突解决（双方功能合并保留）：
- cmd/migrate/main.go：保留 main 的 registerMigrations 重构 + rollback/
  dry-run 子命令；注册表追加 stage 的 CreateApprovalAndSweepTables(023)
  + AddAmlRiskLevelToApprovalRecords(024)，按 version 顺序插在 016 之后
- internal/routes/routes.go：public 组同时保留 main 的 /fund/stats 与
  stage 的 /cosigner/callback（条件注册 + SafeheronRateLimiter）
- internal/container/container.go：保留 main 的 RateLimiter(5/60s) +
  SkipPath("/api/fund/stats") 豁免 + stage 的 SafeheronRateLimiter
  (60/min) 供 cosigner/webhook 使用

review 修复：
- container.go:211 cosigner key path 补 production panic guard，对齐
  同文件 K-16/L-1/COSIGNER_ALLOWED_IPS 模式（原逻辑倒置：路径配了但
  无效会 panic，路径根本没配反而只 log，导致 sweep 审批静默 404）
- routes.go:73,113 注释 10/min → 5/min（merge 把 RateLimiter 从
  10/Minute 收紧到 main 的 5/60s 后注释没同步）
- migrations_test.go TestMigrationOrder 补 023/024 同步 registerMigrations

验证：go build ./cmd/server ./cmd/migrate 通过；approval/container/
safeheron/migrations 测试 ok；handlers 仅 baseline 预存在失败。

+++ internal/repository/postgres/deposit.go
@@ -18,9 +18,14 @@ func NewDepositRepository(db *sql.DB) repository.Deposit {
 }
 
 func (r *DepositRepository) Create(ctx context.Context, deposit *models.Deposit) error {
+	// amount is $3::numeric to bridge the text parameter to the
+	// NUMERIC(32, 8) column installed by migration 047. The implicit
+	// text→numeric cast would also work, but the explicit cast makes
+	// the type contract obvious at the call site and protects against
+	// a future PG upgrade tightening implicit-cast rules.
 	query := `
 		INSERT INTO deposits (user_id, tx_hash, amount, asset, chain, status, from_address, to_address, created_at)
-		VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9)
+		VALUES ($1, $2, $3::numeric, $4, $5, $6, $7, $8, $9)
 		RETURNING id`
 
 	err := r.db.QueryRowContext(ctx, query,
